% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sig_dimensions}
\alias{sig_dimensions}
\title{Get Signature Dimensions}
\usage{
sig_dimensions(d, m)
}
\arguments{
\item{d}{An integer specifying the path dimension.}

\item{m}{An integer specifying the maximum signature order.}
}
\value{
An integer vector of length m, where element k contains the dimension
of the k-th order signature (d^k).
}
\description{
Returns the dimensions of each signature level for given path dimension and order.
}
\examples{
\dontrun{
dims <- sig_dimensions(3, 4)  # c(3, 9, 27, 81)
total_length <- sum(dims)     # 120
}

}
